+++ bst.hpp
@@ -2,6 +2,7 @@
 #define BST_HPP
 
 #include <cstdint>
+#include <cstddef>
 
 struct BSTPrivate;
 
@@ -18,23 +19,23 @@ public:
 	 * Sets the value of a node.
 	 * @returns the number of currently allocated nodes.
 	 */
-	uint32_t set(uint32_t key, uint32_t value);
+	size_t set(uint32_t key, uint32_t value);
 
 	/**
 	 * @returns Sum of values in the interval [left, right].
 	 */
-	uint64_t sum(uint32_t time, uint32_t left, uint32_t right) const;
+	uint64_t sum(size_t time, uint32_t left, uint32_t right) const;
 
 	/**
 	 * Clears the tree at @a time from the history.
 	 * @returns the number of currently allocated nodes.
 	 */
-	uint32_t clear(uint32_t time);
+	size_t clear(size_t time);
 	
 	/**
 	 * @returns the current time.
 	 */
-	uint32_t now() const;
+	size_t now() const;
 
 private:
 	BSTPrivate *d;
